% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Henderson.1988.R
\docType{data}
\name{Henderson.1988}
\alias{Henderson.1988}
\title{Example from Henderson (1988)}
\format{Data frame with 4 rows and 10 columns}
\source{
Henderson CR (1988) Use of an average numerator relationship matrix for multiple-sire joining. Journal of Animal Science 66, 1614-1621.
}
\usage{
Henderson.1988
}
\description{
Pedigree file including diploids of uncertain parentage
}
\references{
Hamilton MG, Kerr RJ. Computation of the inverse additive relationship matrix for autopolyploid and multiple-ploidy populations
}
\author{
Matthew Hamilton <matthew.hamilton@csiro.au>
}
\keyword{datasets}
