% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OR4.1.Ham.Kerr.R
\docType{data}
\name{OR4.1.Ham.Kerr}
\alias{OR4.1.Ham.Kerr}
\title{Example from Hamilton and Kerr}
\format{Data frame with 8 rows and 10 columns}
\source{
Hamilton MG, Kerr RJ. Computation of the inverse additive relationship matrix for autopolyploid and multiple-ploidy populations
}
\usage{
OR4.1.Ham.Kerr
}
\description{
Pedigree file including diploid and triploid pedigree with normal meiosis and second division restitution.
}
\references{
Hamilton MG, Kerr RJ. Computation of the inverse additive relationship matrix for autopolyploid and multiple-ploidy populations
}
\author{
Matthew Hamilton <matthew.hamilton@csiro.au>
}
\keyword{datasets}
