% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OR4.1.Ham.Kerr.uncert.R
\docType{data}
\name{OR4.1.Ham.Kerr.uncert}
\alias{OR4.1.Ham.Kerr.uncert}
\title{Example from Hamilton and Kerr 2018}
\format{Data frame with 8 rows and 11 columns}
\source{
Hamilton MG, Kerr RJ (2018) Computation of the inverse additive relationship matrix for autopolyploid and multiple-ploidy populations. 131:851-860. doi: 10.1007/s00122-017-3041-y
}
\usage{
OR4.1.Ham.Kerr.uncert
}
\description{
Pedigree file including diploid and tetraploid pedigree with uncertain pedigree, normal meiosis, first division restitution and somatic chromosome doubling.
}
\references{
Hamilton MG, Kerr RJ (2018) Computation of the inverse additive relationship matrix for autopolyploid and multiple-ploidy populations. 131:851-860. doi: 10.1007/s00122-017-3041-y
}
\author{
Matthew Hamilton <matthew.hamilton@csiro.au>
}
\keyword{datasets}
