% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polyAinv.R
\name{polyAinv}
\alias{polyAinv}
\title{polyAinv}
\usage{
polyAinv(ped, ASReml.giv.only = FALSE)
}
\arguments{
\item{ped}{A 3-, 4-, 7- or 8-column dataframe containing data of class numeric and/or integer.}

\item{ASReml.giv.only}{Logical.  If TRUE only ASReml.giv is returned.  This can reduce memory requirements.}
}
\value{
A list with the following objects (Hamilton and Kerr):

- K.parents:  Coancestry (K) matrix including parents only

- K.inv:      Inverse coancestry (K) triangular matrix as a 3-column (triplet) dataframe with zero elements removed

- A.parents:  Numerator relationship (A) matrix including parents only

- A.inv:      Inverse numerator relationship (A) triangular matrix as a 3-column (triplet) dataframe with zero elements removed

- ASReml.giv: A.inv as an ASReml-R 'ginverse' object (see 'asreml.Ainverse' in Butler et al. 2009 and '.giv' file in Gilmour et al. 2014)

- F:          Dataframe with 5 columns: INDIV.ID (Individual identifier), PLOIDY (Individual ploidy level, e.g. 2 for diploids), A.DIAG (diagonal elements of the A matrix), K.DIAG (diagonal elements of the K matrix), F (Coefficient of inbreeding)
}
\description{
Computes the additive relationship matrix (A), A inverse,
coancestry matrix (K), K inverse and inbreeding coefficients (F) for autopolyploid and multiple-ploidy populations
(Hamilton and Kerr), and populations in which individuals have uncertain pedigree (Henderson 1988).
}
\details{
Columns in the 3-, 4-, 7- or 8-column 'ped' dataframe must be in the following order:

- Individual identifier (required)

- Sire identifier (required)

- Dam identifier (required)

- Sire gamete ploidy level (lowercase tau in Hamilton and Kerr; required in 7- and 8-column dataframes only)

- Sire lambda - the probability that two genes, drawn at random from a random locus in the sire gamete, are identical
by decent as a result of two copies of a gene being inherited from a single parental chromosome.  This may
be greater than zero in non-monoploid gametes through double reduction, irregular meiosis (e.g. the formation
of unreduced gametes through first division restitution or second division restitution), double fertilisation,
or irregular somatic chromosome doubling. (Hamilton and Kerr; required in 7- and 8-column dataframes only)

- Dam gamete ploidy level (lowercase tau in Hamilton and Kerr; required in 7- and 8-column dataframes only)

- Dam lambda - the probability that two genes, drawn at random from a random locus in the dam gamete, are identical
by decent as a result of two copies of a gene being inherited from a single parental chromosome.  This may
be greater than zero in non-monoploid gametes for the reasons outlined for 'Sire lambda' above (Hamilton and Kerr;
required in 7- and 8-column dataframes only)

- Probability that the individual is the progeny of the sire and dam (Henderson 1988; required in 4- and 8-column dataframes only)

Note:

- column headings in 'ped' are ignored

- if 'ped' contains 3-columns it is assumed to be a standard diploid pedigree file

- if 'ped' contains 4-columns it is assumed to be a diploid pedigree file in which individuals have uncertain pedigree (Henderson 1988)

- if 'ped' contains 7-columns it is assumed to be a polyploid pedigree (Hamilton and Kerr)

- if 'ped' contains 8-columns it is assumed to be a polyploid pedigree in which individuals have uncertain pedigree (Hamilton and Kerr; Henderson 1988)
}
\examples{
#7-column example with diploids, tetraploids, first division restitution,
#second division restitution and parthenogenesis
data(Tab.1.Ham.Kerr)
polyAinv(ped = Tab.1.Ham.Kerr[,1:7])

#7-column example with diploids, triploids and second division restitution
data(OR4.1.Ham.Kerr)
polyAinv(ped = OR4.1.Ham.Kerr[,1:7])

#7-column example with diploids, tetraploids, first division restitution and
#somatic chromosome doubling.  Returns an ASReml ginverse object only.
data(OR4.6.Ham.Kerr)
polyAinv(ped = OR4.6.Ham.Kerr[,1:7], ASReml.giv.only = TRUE)

#4-column example from Henderson (1988)
data(Henderson.1988)
polyAinv(ped = Henderson.1988[,1:4])

#8-column example (OR4.1.Ham.Kerr example pedigree modified)
data(OR4.1.Ham.Kerr.uncert)
polyAinv(ped = OR4.1.Ham.Kerr.uncert[,c(1:7,11)])

#All tetraploid example with double reduction in founders (Tab.4.Kerr.et.al.2012).  
#Note that results differ from Table 4 of Kerr et al. (2012) and those using the 
#approach of Amadeu (2016), as inbreeding in founders is accounted for.
data(Tab.4.Kerr.et.al.2012)
polyAinv(ped = Tab.4.Kerr.et.al.2012[,c(1:7)])
}
\references{
Amadeu RR, Cellon C, Olmstead JW, Garcia AAF, Resende MFR, Munoz PR (2016) AGHmatrix: R package to construct relationship matrices for autotetraploid and diploid species. A blueberry example. The Plant Genome 9

Butler DG, Cullis BR, Gilmour AR, Gogel BJ (2009) ASReml-R reference manual. Queensland Department of Primary Industries and Fisheries, Brisbane

Gilmour AR, Gogel BJ, Cullis BR, Welham SJ, Thompson R (2014) ASReml user guide release 4.1 functional specification. VSN International Ltd, Hemel Hempstead, UK

Hamilton MG, Kerr RJ. (in press) Computation of the inverse additive relationship matrix for autopolyploid and multiple-ploidy populations.  Theoretical and Applied Genetics

Henderson CR (1988) Use of an average numerator relationship matrix for multiple-sire joining. Journal of Animal Science 66, 1614-1621.
}
\author{
Matthew Hamilton <matthew.hamilton@csiro.au>
}
