% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Tab.4.Kerr.et.al.2012.R
\docType{data}
\name{Tab.4.Kerr.et.al.2012}
\alias{Tab.4.Kerr.et.al.2012}
\title{Example from Kerr et al 2012}
\format{Data frame with 8 rows and 10 columns}
\source{
Kerr RJ, Li L, Tier B, Dutkowski GW, McRae TA (2012) Use of the numerator relationship matrix in genetic analysis of autopolyploid species. Theor Appl Genet 124:1271-1282. doi: DOI 10.1007/s00122-012-1785-y
}
\usage{
Tab.4.Kerr.et.al.2012
}
\description{
Pedigree file for tetraploid with double reduction in Table 4
}
\references{
Hamilton MG, Kerr RJ (2018) Computation of the inverse additive relationship matrix for autopolyploid and multiple-ploidy populations. 131:851-860. doi: 10.1007/s00122-017-3041-y
}
\author{
Matthew Hamilton <matthew.hamilton@csiro.au>
}
\keyword{datasets}
